# Conflicts:
#	features/content/feed/src/main/java/com/tokopedia/feedplus/view/adapter/typefactory/dynamicfeed/DynamicFeedTypeFactoryImpl.kt
#	features/content/feed/src/main/java/com/tokopedia/feedplus/view/presenter/FeedViewModel.kt

+++ features/content/kol/src/main/java/com/tokopedia/kol/feature/comment/view/presenter/KolCommentPresenter.java
@@ -16,7 +16,6 @@ import com.tokopedia.kol.feature.comment.view.subscriber.ReportKolSubscriber;
 import com.tokopedia.kol.feature.comment.view.subscriber.SendKolCommentSubscriber;
 import com.tokopedia.kol.feature.report.domain.usecase.SendReportUseCase;
 
-import java.lang.ref.WeakReference;
 import java.util.List;
 
 import javax.inject.Inject;
@@ -62,7 +61,7 @@ public class KolCommentPresenter extends BaseDaggerPresenter<KolComment.View>
     }
 
     @Override
-    public void getCommentFirstTime(int postId) {
+    public void getCommentFirstTime(long postId) {
         getView().showLoading();
         getKolCommentsUseCase.execute(
                 GetKolCommentsUseCase.getFirstTimeParam(postId),
@@ -70,7 +69,7 @@ public class KolCommentPresenter extends BaseDaggerPresenter<KolComment.View>
     }
 
     @Override
-    public void loadMoreComments(int postId) {
+    public void loadMoreComments(long postId) {
         getKolCommentsUseCase.execute(
                 GetKolCommentsUseCase.getParam(postId, cursor),
                 new GetKolCommentSubscriber(getView()));
@@ -85,7 +84,7 @@ public class KolCommentPresenter extends BaseDaggerPresenter<KolComment.View>
     public void deleteComment(String id, int adapterPosition) {
         getView().showProgressDialog();
         deleteKolCommentUseCase.execute(
-                DeleteKolCommentUseCase.getParam(Integer.parseInt(id)),
+                DeleteKolCommentUseCase.getParam(Long.parseLong(id)),
                 new DeleteKolCommentSubscriber(getView(), adapterPosition)
         );
     }
@@ -99,7 +98,7 @@ public class KolCommentPresenter extends BaseDaggerPresenter<KolComment.View>
     }
 
     @Override
-    public void sendComment(int id, String comment) {
+    public void sendComment(long id, String comment) {
         if (isValid(comment)) {
             getView().showProgressDialog();
             getView().disableSendComment();
